chore: consolidate global and context hooks
