+++ src/components/LeaderBoard.js
@@ -1,17 +1,10 @@
-import {
-  Paper,
-  Box,
-  Grid,
-  Typography,
-  useMediaQuery,
-  useTheme,
-  SvgIcon,
-} from '@mui/material';
+import { Paper, Box, Grid, Typography, SvgIcon } from '@mui/material';
 import countries from 'i18n-iso-countries';
 import Image from 'next/image';
 import { makeStyles } from 'models/makeStyles';
 import { fixCountryNames } from 'models/utils';
 import Ribbon from './Ribbon';
+import { useMobile } from '../hooks/globalHooks';
 import TreeIcon from '../images/icons/tree.svg';
 
 const useStyles = makeStyles()((theme) => ({
@@ -96,15 +89,13 @@ function RibbonWrapper({ fill, index }) {
   );
 }
 
-function TreeImage() {
-  const theme = useTheme();
-  const isMobileScreen = useMediaQuery(theme.breakpoints.down('md'));
+function TreeImage(isMobile) {
   return (
     <SvgIcon
       component={TreeIcon}
       inheritViewBox
       sx={`${
-        !isMobileScreen
+        !isMobile
           ? 'width: 13.5px; height: 18px;'
           : 'width: 12px; height: 14px;'
       }`}
@@ -115,8 +106,7 @@ function TreeImage() {
 
 function LeaderBoard(props) {
   const { countries: rankedCountries, handleCountryClick } = props;
-  const theme = useTheme();
-  const isMobileScreen = useMediaQuery(theme.breakpoints.down('md'));
+  const isMobile = useMobile();
 
   const fixedCountries = fixCountryNames(rankedCountries);
 
@@ -182,7 +172,7 @@ function LeaderBoard(props) {
             >
               TREES PLANTED
             </Typography>
-            <TreeImage />
+            <TreeImage isMobile={isMobile} />
           </Grid>
           <Grid item xs={1} />
         </Grid>
@@ -193,7 +183,7 @@ function LeaderBoard(props) {
           <Paper
             key={country.id}
             onClick={() => handleCountryClick(country.id)}
-            elevation={isMobileScreen ? 2 : 5}
+            elevation={isMobile ? 2 : 5}
             sx={{
               borderRadius: '100px',
               // boxShadow: '0 4px 10px 0px rgba(0, 0, 0, 0.15)',
