+++ src/components/PlanterQuote.js
@@ -1,19 +1,11 @@
- 
-import {
-  Avatar,
-  Button,
-  Box,
-  Grid,
-  Typography,
-  useTheme,
-  SvgIcon,
-} from '@mui/material';
+import { Avatar, Button, Box, Grid, Typography, SvgIcon } from '@mui/material';
 import * as d3 from 'd3';
 import moment from 'moment';
 import Image from 'next/image';
 import ColorButton from './common/ColorButton';
 import DataTag from './common/DataTag';
 import Info from './common/Info';
+import { useMobile } from '../hooks/globalHooks';
 import CalendarIcon from '../images/icons/calendar.svg';
 import LocationIcon from '../images/icons/location.svg';
 import PeopleIcon from '../images/icons/people.svg';
@@ -30,8 +22,7 @@ function PlanterQuote({
   location,
   reverse = false,
 }) {
-  const theme = useTheme();
-  const isMobile = theme.breakpoints.down('sm');
+  const isMobile = useMobile();
   return (
     <Box
       sx={{
