+++ src/components/playground/FontAddInput.js
@@ -11,7 +11,7 @@ import {
   InputAdornment,
 } from '@mui/material';
 import { useState } from 'react';
-import { usePlaygroundFonts } from '../../context/playgroundContext';
+import { usePlaygroundFonts } from '../../hooks/contextHooks';
 import { loadFonts } from '../../models/utils';
 
 function FontAddInput() {
