+++ src/components/playground/FontCustomization.js
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import FontAddInput from './FontAddInput';
 import FontsList from './FontsList';
-import { usePlaygroundFonts } from '../../context/playgroundContext';
+import { usePlaygroundFonts } from '../../hooks/contextHooks';
 
 const popularFonts = [
   'Lora',
