+++ src/components/playground/ToggleThemeMode.js
@@ -6,7 +6,7 @@ import {
   Switch,
 } from '@mui/material';
 import { useCallback, useEffect, useState } from 'react';
-import { usePlaygroundThemeType } from '../../context/playgroundContext';
+import { usePlaygroundThemeType } from '../../hooks/contextHooks';
 
 function ToggleThemeMode() {
   const [themeType, setThemeType] = usePlaygroundThemeType();
