+++ src/pages/_app.js
@@ -2,7 +2,6 @@ import '../style.css';
 
 import createCache from '@emotion/cache';
 import { CacheProvider } from '@emotion/react';
-import { useMediaQuery, useTheme } from '@mui/material';
 import log from 'loglevel';
 import { useRouter } from 'next/router';
 import Layout from '../components/Layout';
@@ -13,7 +12,7 @@ import LayoutMobileB from '../components/LayoutMobileB';
 import LayoutMobileC from '../components/LayoutMobileC';
 import { DrawerProvider } from '../context/DrawerContext';
 import { CustomThemeProvider } from '../context/themeContext';
-import useEmbed from '../hooks/useEmbed';
+import { useMobile, useEmbed } from '../hooks/globalHooks';
 import { MapContextProvider } from '../mapContext';
 
 if (process.env.NEXT_PUBLIC_API_MOCKING === 'enabled') {
@@ -32,8 +31,7 @@ export const createMuiCache = () =>
   }));
 
 function TreetrackerApp({ Component, pageProps }) {
-  const theme = useTheme();
-  const nextExtraIsDesktop = useMediaQuery(theme.breakpoints.up('sm'));
+  const nextExtraIsDesktop = !useMobile();
   const nextExtraIsEmbed = useEmbed();
   log.warn('app: isDesktop: ', nextExtraIsDesktop);
   log.warn('app: component: ', Component);
